add logic to deal with loading status and refactoring
