+++ public/pages/Dashboard/Components/utils/DashboardHeader.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { EuiPageHeader, EuiTitle } from '@elastic/eui';
+export const DashboardHeader = () => {
+  return (
+    <EuiPageHeader>
+      <EuiTitle size="l">
+        <h1>Dashboard</h1>
+      </EuiTitle>
+    </EuiPageHeader>
+  );
+};
